requirements.txt
wordcloud
Pillow
pandas
matplotlib
numpy
google-play-scraper == 0.0.2.0
